+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -22,6 +22,7 @@
 
 #### Other Changes
 * Added `x-ms-hub-region-processing-only` header to allow hub-region stickiness when 404 `READ SESSION NOT AVAIALBLE` is hit for Single-Writer accounts. - [PR 47631](https://github.com/Azure/azure-sdk-for-java/pull/47631)
+* Added aggressive HTTP timeout policies for document operations routed to Gateway V2. - [PR 47879](https://github.com/Azure/azure-sdk-for-java/pull/47879)
 
 ### 4.77.0 (2026-01-26)